requirements.txt

langchain
langchain-community
langchain-core
ollama
fastapi
uvicorn
requests
sentence-transformers
scikit-learn
numpy
chromadb
pydantic_settings
unstructured
